+++ Changelog.md
@@ -6,6 +6,8 @@
 
 - **Added** ability to customize default neighbor expansion limit
   ([#925](https://github.com/aws/graph-explorer/pull/925))
+- **Added** ability to resize the sidebar
+  ([#938](https://github.com/aws/graph-explorer/pull/938))
 - **Improved** Improved connection details by moving the sync button near the
   last sync timestamp (thanks @enumura1,
   [#901](https://github.com/aws/graph-explorer/pull/901),